requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
jinja2==3.1.4
sqlmodel==0.0.21
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pillow==10.4.0
qrcode==7.4.2
python-multipart==0.0.9
